+++ src/features/watch/WatchPage.tsx
@@ -2,29 +2,50 @@ import { Link, useLocation } from "react-router";
 import { useQueueHandler } from "./QueueHandler";
 import useLocalStorageState from "../../hooks/useLocalStorage";
 import data from "../../data/db";
-import type { Review } from "../../types/types";
+import type { Playlist, Rating, Short } from "../../types/types";
+import { useState } from "react";
+import { buildSessionShorts } from "../../lib/session";
 
 export default function WatchPage() {
   const { playlistId } = useLocation().state as { playlistId: string };
-  const [playlists, setPlaylists] = useLocalStorageState("playlists", data);
+  const [playlists, setPlaylists] = useLocalStorageState<Playlist[]>(
+    "playlists",
+    data,
+  );
 
-  const shorts = playlists.find((p) => p.id === playlistId)?.shorts ?? [];
+  const [selected, setSelected] = useState<string | null>(null);
 
-  function onComplete(review: Review) {
-    console.log("session finished");
-    console.log(review);
+  function handleClick(rating: Rating) {
+    respond(rating);
+    setSelected(rating);
+    setTimeout(() => setSelected(null), 400);
+  }
 
+  const [sessionStartedAt] = useState(() => Date.now());
+
+  const playlist = playlists.find((p) => p.id === playlistId);
+  const allShorts = playlist?.shorts ?? [];
+  const settings = playlist?.settings ?? { newLimit: 5, reviewLimit: 5 };
+
+  const dueShorts = buildSessionShorts(allShorts, settings, sessionStartedAt);
+
+  function onComplete(playlistId: string, updatedShorts: Short[]) {
     setPlaylists((prev) =>
       prev.map((playlist) =>
-        playlist.id === review.playlistId
-          ? { ...playlist, watchCount: playlist.watchCount + 1 }
+        playlist.id === playlistId
+          ? {
+              ...playlist,
+              watchCount: playlist.watchCount + 1,
+              shorts: updatedShorts,
+            }
           : playlist,
       ),
     );
   }
 
   const { currentShort, respond, done } = useQueueHandler(
-    shorts,
+    dueShorts,
+    allShorts,
     playlistId,
     onComplete,
   );
@@ -44,7 +65,7 @@ export default function WatchPage() {
     <div className="watch-container flex flex-col flex-1 md:mb-12 md:mx-24">
       <div className="screen border flex-1">
         <iframe
-          src={`https://www.youtube.com/embed/${currentShort}`}
+          src={`https://www.youtube.com/embed/${currentShort.id}`}
           className="w-full h-full"
           allowFullScreen
         />
@@ -58,28 +79,32 @@ export default function WatchPage() {
           Exit
         </Link>
         <button
-          className="again rounded flex-1 py-2 uppercase text-2xl tracking-widest bg-red-300"
-          onClick={() => respond("again")}
+          className={`easy rounded flex-1 py-2 uppercase text-2xl tracking-widest transition
+    ${selected === "again" ? "bg-black text-white" : "bg-red-300"}`}
+          onClick={() => handleClick("again")}
         >
-          Again
+          {selected === "again" ? "✔" : "Again"}
         </button>
         <button
-          className="hard rounded flex-1 py-2 uppercase text-2xl tracking-widest bg-orange-300"
-          onClick={() => respond("hard")}
+          className={`easy rounded flex-1 py-2 uppercase text-2xl tracking-widest transition
+    ${selected === "hard" ? "bg-black text-white" : "bg-orange-300"}`}
+          onClick={() => handleClick("hard")}
         >
-          Hard
+          {selected === "hard" ? "✔" : "Hard"}
         </button>
         <button
-          className="medium rounded flex-1 py-2 uppercase text-2xl tracking-widest bg-green-300"
-          onClick={() => respond("medium")}
+          className={`easy rounded flex-1 py-2 uppercase text-2xl tracking-widest transition
+    ${selected === "medium" ? "bg-black text-white" : "bg-green-300"}`}
+          onClick={() => handleClick("medium")}
         >
-          OK
+          {selected === "medium" ? "✔" : "Ok"}
         </button>
         <button
-          className="easy rounded flex-1 py-2 uppercase text-2xl tracking-widest bg-blue-300"
-          onClick={() => respond("easy")}
+          className={`easy rounded flex-1 py-2 uppercase text-2xl tracking-widest transition
+    ${selected === "easy" ? "bg-black text-white" : "bg-blue-300"}`}
+          onClick={() => handleClick("easy")}
         >
-          Easy
+          {selected === "easy" ? "✔" : "Easy"}
         </button>
       </div>
     </div>